backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
python-dotenv
webdriver-manager

